// IDDD_AgilePM_Example_01/src/domain/model/agilepm/team/MemberType.ts
//   Copyright 2012-2026 Vaughn Vernon. All rights reserved.
//
//   Licensed under the Apache License, Version 2.0 (the "License");
//   you may not use this file except in compliance with the License.
//   You may obtain a copy of the License at
//
//       http://www.apache.org/licenses/LICENSE-2.0
//
//   Unless required by applicable law or agreed to in writing, software
//   distributed under the License is distributed on an "AS IS" BASIS,
//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//   See the License for the specific language governing permissions and
//   limitations under the License.

/**
 * Distinguishes between ProductOwner and TeamMember within a Team.
 */
export enum MemberType {
    ProductOwner = 'ProductOwner',
    TeamMember = 'TeamMember'
}